+++ src/main/java/io/github/astrapi69/swing/tree/content/panel/GenericTreeNodeWithContentPanel.java
@@ -31,7 +31,8 @@ import io.github.astrapi69.model.api.Model;
 import io.github.astrapi69.tree.TreeNode;
 
 public abstract class GenericTreeNodeWithContentPanel<T, C>
-	extends JXTreeWithContentPanel<TreeNode<T>, C>
+	extends
+		JXTreeWithContentPanel<TreeNode<T>, C>
 {
 
 	/**