+++ src/main/java/io/github/astrapi69/swing/tree/panel/node/NodePanel.java
@@ -28,17 +28,17 @@ import java.awt.*;
 
 import javax.swing.*;
 
-import io.github.astrapi69.model.LambdaModel;
-import io.github.astrapi69.swing.JMCheckBox;
-import io.github.astrapi69.swing.JMTextField;
 import lombok.Getter;
-import io.github.astrapi69.model.BaseModel;
-import io.github.astrapi69.model.api.Model;
-import io.github.astrapi69.swing.base.BasePanel;
 import net.miginfocom.layout.AC;
 import net.miginfocom.layout.CC;
 import net.miginfocom.layout.LC;
 import net.miginfocom.swing.MigLayout;
+import io.github.astrapi69.model.BaseModel;
+import io.github.astrapi69.model.LambdaModel;
+import io.github.astrapi69.model.api.Model;
+import io.github.astrapi69.swing.JMCheckBox;
+import io.github.astrapi69.swing.JMTextField;
+import io.github.astrapi69.swing.base.BasePanel;
 
 @Getter
 public class NodePanel extends BasePanel<NodeModelBean>
@@ -94,10 +94,8 @@ public class NodePanel extends BasePanel<NodeModelBean>
 
 	protected void onInitializeMigLayout()
 	{
-		MigLayout layout = new MigLayout(
-			new LC().fillX().wrapAfter(2),
-			new AC().align("left").gap("10").grow().fill(),
-			new AC().fill().gap("10"));
+		MigLayout layout = new MigLayout(new LC().fillX().wrapAfter(2),
+			new AC().align("left").gap("10").grow().fill(), new AC().fill().gap("10"));
 		this.setLayout(layout);
 
 		add(lblName);